+++ app/Modules/Hevy/Models/HevyWorkoutExercise.php
@@ -2,6 +2,7 @@
 
 namespace App\Modules\Hevy\Models;
 
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
 class HevyWorkoutExercise extends Model